feat: refactor cli structure implement daemon command
