+++ src/cli/cleanup.ts
